+++ src/main/java/kg/attractor/jobsearch_remake/repository/MessageRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface MessageRepository extends JpaRepository<Message, Integer> {
-    List<Message> findByRespondedApplicants(Integer respondedApplicantId);
+    List<Message> findByRespondedApplicants(Long respondedApplicantId);
 }
